youtube-summarizer-AI/requirements.txt
streamlit
python-dotenv
groq
yt-dlp
youtube-transcript-api
reportlab
pypdf